requirements.txt
numpy>=1.24
matplotlib>=3.7
transformers
torch
pillow
requests
